Add BOM to project file and remove local.settings.json block
Added a Unicode Byte Order Mark (BOM) to the beginning of the `inprocess-net6-openai.csproj` file to indicate the encoding of the text file. Also, removed the `<None Update="local.settings.json">` block from the same file, which may change the default behavior for the `local.settings.json` file during the build process.